requirements.txt
# 核心框架
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0

# 向量数据库
chromadb>=0.5.0,<0.6.0

# 文档处理
pypdf>=4.0.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0

# Web界面
streamlit>=1.35.0,<2.0.0
pandas>=2.0.0,<3.0.0
openpyxl>=3.1.0,<4.0.0

# AI API
zhipuai>=2.1.0
openai>=1.30.0,<2.0.0

# 工具
requests>=2.31.0,<3.0.0
pyyaml>=6.0,<7.0